# [Draft] 2x performance of `terrain_fill_connect`