[draft, gonna clean it up a bit]

A conglomeration of small tweaks found via XCode Instruments. Further improvements might require a bigger rewrite.

Before: **~350ms on average** (as measured in my test harness)
After: **~165ms on average**, 2.1x faster 

Moving to HashSet over RBSet: **saves ~95ms**
Removing unnecessary RBMap for minimization: **saes ~40ms**
Moving get_terrains_pattern_set to parent: **saves ~50ms** 

Sample test code:

```
extends TileMapLayer

var toggle = false

func _process(delta):
  var startTime = Time.get_ticks_msec()
  var points = []
  for i in range(200):
    for j in range(200):
      points.append(
        Vector2i(i, j)
      )
  var allocTime = Time.get_ticks_msec()
  if toggle:
    set_cells_terrain_connect(points, 0, -1, false)
  else:
    set_cells_terrain_connect(points, 0, 0, false)
  toggle = !toggle
  var totalTime = Time.get_ticks_msec()
  print("Alloc time:", allocTime - startTime)
  print("Terrain Calc time:", totalTime - allocTime)
```

Thanks for making it so easy to contribute that it's as easy to fix the underlying issue as it is to work around it in my game :)

<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://contributing.godotengine.org/en/latest/engine/guidelines/code_style.html
-->